consumer/app/requirements.txt
confluent-kafka==2.1.0
kafka-python==2.0.2
prometheus-client==0.21.1